requirements.txt
python-magic
pyclamd
yara-python
pdfid
bleach
frictionless
vt-py
zstandard
pandas
pyarrow
python-dotenv
PyYAML
